[refactor] 3단계 - 자동차 경주 테스트코드 코드리뷰 반영
1. 필드명 변경
2. 패키지명 소문자 및 숫자 컨벤션 수정
3. 테스트코드 클래스명따라서 분리
4. 상수 위치 컨벤션 수정
5. input 클래스 수정
