+++ src/main/java/step3racingcarcompetition/Car.java
@@ -1,18 +1,18 @@
-package step3RacingCarCompetition;
+package step3racingcarcompetition;
 
 public class Car {
-    private String movingData = "";
+    private String position = "";
 
     public void recordCarData(int totalMovingCount, RandomGenerator randomGenerator) {
         StringBuilder movingDataBuilder = new StringBuilder();
         for (int movingCount = 0; movingCount < totalMovingCount; movingCount++) {
             movingDataBuilder.append(randomGenerator.generateRandom());
         }
-        movingData = movingDataBuilder.toString();
+        position = movingDataBuilder.toString();
     }
 
     public String findEachMovingData(int movingCount) {
-        return movingData.substring(0, movingCount);
+        return position.substring(0, movingCount);
     }
 
 }
