+++ src/main/java/step3racingcarcompetition/Main.java
@@ -1,4 +1,4 @@
-package step3RacingCarCompetition;
+package step3racingcarcompetition;
 
 public class Main {
     public static void main(String[] args) {
