+++ src/main/java/step3racingcarcompetition/RacingCarInfo.java
@@ -1,4 +1,4 @@
-package step3RacingCarCompetition;
+package step3racingcarcompetition;
 
 import java.util.ArrayList;
 import java.util.List;
