+++ src/main/java/step3racingcarcompetition/RandomGenerator.java
@@ -1,5 +1,6 @@
-package step3RacingCarCompetition;
+package step3racingcarcompetition;
 
+@FunctionalInterface
 public interface RandomGenerator {
     int generateRandom();
 }
